+++ tests/resources/TestModules/WriteHello/public/Write-Hello.ps1
@@ -1,9 +1,21 @@
+<#
+.DESCRIPTION
+    A simple function that writes a greeting to the console
+.EXAMPLE
+    Write-Hello
+    Hello, World!
+.PARAMETER Name
+    The name of the person to greet
+#>
+
 function Write-Hello {
     [CmdletBinding()]
     param (
+        # The name of the person to greet
         [Parameter()]
-        $Name = 'World',
+        [string]$Name = 'World',
 
+        # The planet to welcome the person to
         [Parameter(Mandatory)]
         [Dynamic({
             $null -eq $PSBoundParameters['Name']