+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.1] Add comment based help support
+
+- Added support for comment based help to be used in the dynamic function definition
+
 ## [1.1.0] House cleaning
 
 - Added use cases to test coverage